requirements-lock.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pyhd8ed1ab_0
asttokens=2.4.1=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py310h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.32.3=h4bc722e_0
ca-certificates=2024.8.30=hbcca054_0
certifi=2024.8.30=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cuda=11.6.1=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=12.4.127=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.4.127=0
cuda-nsight-compute=12.4.1=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=12.4.127=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=12.4.127=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=12.4.127=0
cuda-runtime=11.6.1=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=12.4.127=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-visual-tools=11.6.1=0
decorator=5.1.1=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.1.0=pypi_0
ffmpeg=4.3=hf484d3e_0
flake8=7.1.1=pypi_0
freetype=2.12.1=h4a9f257_0
gds-tools=1.9.1.3=0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpcio=1.62.2=py310h6a678d5_0
idna=3.7=py310h06a4308_0
importlib-metadata=8.4.0=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipython=8.27.0=pypi_0
jedi=0.19.1=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libcublas=11.9.2.110=h5e84587_0
libcublas-dev=11.9.2.110=h5c901ab_0
libcufft=10.7.1.112=hf425ae0_0
libcufft-dev=10.7.1.112=ha5ce4c0_0
libcufile=1.9.1.3=0
libcufile-dev=1.9.1.3=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcusolver=11.3.4.124=h33c3c4e_0
libcusparse=11.7.2.124=h7538f96_0
libcusparse-dev=11.7.2.124=hbbe9722_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc=14.1.0=h77fa898_1
libgcc-ng=14.1.0=h69a702a_1
libgomp=14.1.0=h77fa898_1
libgrpc=1.62.2=h2d74bed_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libnpp=11.6.3.124=hd2722f0_0
libnpp-dev=11.6.3.124=h3c42840_0
libnvjpeg=11.6.2.124=hd473ad6_0
libnvjpeg-dev=11.6.2.124=hb5906b9_0
libpng=1.6.39=h5eee18b_0
libprotobuf=4.25.3=he621ea3_0
libstdcxx=14.1.0=hc0a3c3a_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.6=pyhd8ed1ab_0
markupsafe=2.1.5=py310ha75aee5_1
matplotlib-inline=0.1.7=pypi_0
mccabe=0.7.0=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.10=py310h5eee18b_0
mkl_random=1.2.7=py310h1128e8f_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
nsight-compute=2024.1.1.4=0
numpy=1.26.4=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.1=pyhd8ed1ab_0
pandas=1.5.3=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.4.0=py310h5eee18b_0
pip=24.2=py310h06a4308_0
prompt-toolkit=3.0.47=pypi_0
protobuf=5.28.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=17.0.0=pypi_0
pycodestyle=2.12.1=pypi_0
pyflakes=3.2.0=pypi_0
pygments=2.18.0=pypi_0
pylance=0.17.0=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.14=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.10=2_cp310
pytorch=1.13.1=py3.10_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
requests=2.32.3=py310h06a4308_0
ruff=0.6.4=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.14.1=pypi_0
setuptools=72.1.0=py310h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.17.1=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py310h75e40e8_1
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
torch-tb-profiler=0.4.3=pypi_0
torchaudio=0.13.1=py310_cu116
torchvision=0.14.1=py310_cu116
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pypi_0
typing_extensions=4.11.0=py310h06a4308_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=py310h06a4308_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.4=pyhd8ed1ab_0
wheel=0.43.0=py310h06a4308_0
xz=5.4.6=h5eee18b_1
zipp=3.20.1=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
